+++ plugins/player.lua
@@ -2,6 +2,8 @@
 
 Player = {}
 
+local auto_callbacks = {}
+
 
 
 -- ========== Static Methods ==========
@@ -51,6 +53,17 @@ Player.get_from_name = function(name)
 end
 
 
+Player.add_callback = function(callback, id, func, skill, all_damage)
+    table.insert(auto_callbacks, {
+        callback,
+        id,
+        func,
+        skill,
+        all_damage
+    })
+end
+
+
 
 -- ========== Instance Methods ==========
 
@@ -92,6 +105,29 @@ methods_player = {
 }
 
 
+methods_player_callbacks = {
+    
+    onStatRecalc        = function(self, id, func) Player.add_callback("onStatRecalc", id, func) end,
+    onPostStatRecalc    = function(self, id, func) Player.add_callback("onPostStatRecalc", id, func) end,
+    onSkillUse          = function(self, id, func, skill) Player.add_callback("onSkillUse", id, func, skill) end,
+    onBasicUse          = function(self, id, func) Player.add_callback("onBasicUse", id, func) end,
+    onAttack            = function(self, id, func, all_damage) Player.add_callback("onAttack", id, func, nil, all_damage) end,
+    onPostAttack        = function(self, id, func, all_damage) Player.add_callback("onPostAttack", id, func, nil, all_damage) end,
+    onHit               = function(self, id, func, all_damage) Player.add_callback("onHit", id, func, nil, all_damage) end,
+    onKill              = function(self, id, func) Player.add_callback("onKill", id, func) end,
+    onDamaged           = function(self, id, func) Player.add_callback("onDamaged", id, func) end,
+    onDamageBlocked     = function(self, id, func) Player.add_callback("onDamageBlocked", id, func) end,
+    onHeal              = function(self, id, func) Player.add_callback("onHeal", id, func) end,
+    onShieldBreak       = function(self, id, func) Player.add_callback("onShieldBreak", id, func) end,
+    onInteract          = function(self, id, func) Player.add_callback("onInteract", id, func) end,
+    onEquipmentUse      = function(self, id, func) Player.add_callback("onEquipmentUse", id, func) end,
+    onPreStep           = function(self, id, func) Player.add_callback("onPreStep", id, func) end,
+    onPostStep          = function(self, id, func) Player.add_callback("onPostStep", id, func) end,
+    onDraw              = function(self, id, func) Player.add_callback("onDraw", id, func) end
+
+}
+
+
 
 -- ========== Metatables ==========
 
@@ -111,3 +147,31 @@ metatable_player = {
         metatable_instance_gs.__newindex(table, key, value)
     end
 }
+
+
+metatable_player_callbacks = {
+    __index = function(table, key)
+        -- Methods
+        if methods_player_callbacks[key] then
+            return methods_player_callbacks[key]
+        end
+    end
+}
+setmetatable(Player, metatable_player_callbacks)
+
+
+
+-- ========== Initialize ==========
+
+Player.__initialize = function()
+    Callback.add("onGameStart", "RMT.player_addAutoCallbacks", function(self, other, result, args)
+        Alarm.create(function()
+            local player = Player.get_client()
+            if player:exists() then
+                for _, c in ipairs(auto_callbacks) do
+                    player:add_callback(c[1], c[2], c[3], c[4], c[5])
+                end
+            end
+        end, 1)
+    end, true)
+end