+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### v1.1.13
 * Instance : get_data tables and Instance callbacks are no longer instantly deleted for players on-death.
+* Player : Added add_callback, which automatically assigns Instance callbacks to the local player on run start.
 * Stage : clear_rooms now removes environment logs properly.
 * Language : translate_token now returns the input if the associated text cannot be found.